[#2224] Move date utils out from workshops.util
