+++ amy/communityroles/templatetags/communityroles.py
@@ -4,7 +4,7 @@ from django import template
 
 from communityroles.models import CommunityRole
 from workshops.models import Person
-from workshops.util import human_daterange as human_daterange_util
+from workshops.utils.dates import human_daterange
 
 register = template.Library()
 
@@ -19,7 +19,7 @@ def get_community_role(person: Person, role_name: str) -> Optional[CommunityRole
 
 @register.simple_tag
 def community_role_human_dates(community_role: CommunityRole) -> str:
-    result = human_daterange_util(
+    result = human_daterange(
         community_role.start, community_role.end, no_date_right="present"
     )
     return result
